requirements.txt
disqus==0.0.4
dj-database-url==0.3.0
Django==1.9.1
django-ckeditor==5.0.2
gunicorn==19.4.1
psycopg2==2.6.1
wheel==0.24.0
whitenoise==2.0.6
